+++ migrations/0_6to0_7.py
@@ -1,6 +1,7 @@
 from playhouse.migrate import migrate, SqliteMigrator, MySQLMigrator
-from peewee import MySQLDatabase, SqliteDatabase
-from app import database
+from peewee import MySQLDatabase, SqliteDatabase, \
+    IntegerField, DateTimeField, ForeignKeyField
+from app import database, User
 
 if type(database) == MySQLDatabase:
     migrator = MySQLMigrator(database)
@@ -11,7 +12,9 @@ else:
     exit()
 
 with database.atomic():
+    database.create_tables([User])
     migrate(
-        
+        migrator.add_column('page', 'calendar', DateTimeField(index=True, null=True)),
+        migrator.add_column('page', 'owner_id', IntegerField(null=True))
     )